chore: remove init and use it directly in the main program; clean files
